requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Click==7.0
dj-database-url==0.5.0
Django==3.0.3
django-heroku==0.3.1
first==2.0.2
geographiclib==1.50
geopy==1.21.0
gunicorn==20.0.4
idna==2.8
pip-tools==1.8.0
psycopg2==2.8.4
python-dotenv==0.12.0
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
whitenoise==5.0.1
